requirements.txt
tensorflow
opencv-python
matplotlib
numpy
requests
pycocotools
Pillow
shapely
